chore(discover): deprecate legacy discovery mode and modernize the route module
DISCOVERY_MODE=legacy is now deprecated: each process logs one startup
warning, the environment reference marks the value deprecated, and removal
is planned for a future release. Legacy behavior is unchanged.

The 2,429-line discover.ts splits into routes/discover/: a thin router
dispatching each endpoint on the mode flag, typed modern handlers (Prisma
and queue boundaries no longer any-typed), and frozen legacy modules moved
verbatim under discover/legacy/ with no fixes applied. The known legacy
DiscoveryBatch duplicate-retry hazard stays tracked in the modernization
roadmap and is retired with the mode.

File-size baseline entry for discover.ts removed; error-canon budgets
rekeyed to the new layout with the same totals.

Closes #597
