+++ docs/ENVIRONMENT_VARIABLES.md
@@ -219,7 +219,7 @@ See the [`OIDC_SSO.md` topology matrix](OIDC_SSO.md#deployment-topology) for pro
 | `FANART_API_KEY`                         | `backend`                                      | Optional                                      | unset                                                                              | Fanart.tv API key for artist images/backgrounds. A database-stored key (System Settings) takes precedence; this var is the `.env` fallback used only when no key is stored. By default, the background enrichment cycle's `ImageProviderService` retains its historical env-only behavior. Under `SECRETS_DB_ONLY=true`, that enrichment path instead reads the stored System Settings key and ignores this env var. |
 | `OPENAI_API_KEY`                         | `backend`, `backend-worker`                    | Optional                                      | unset                                                                              | OpenAI key for AI-assisted recommendation features.                                                                                                                                                                                                                                                                                                                                                                  |
 | `DEEZER_API_KEY`                         | `backend`, `backend-worker`                    | Optional                                      | unset                                                                              | Deezer API key override.                                                                                                                                                                                                                                                                                                                                                                                             |
-| `DISCOVERY_MODE`                         | `backend`, `backend-worker`                    | Optional                                      | `recommendation`                                                                   | Discovery mode (`recommendation` or `legacy`).                                                                                                                                                                                                                                                                                                                                                                       |
+| `DISCOVERY_MODE`                         | `backend`, `backend-worker`                    | Optional                                      | `recommendation`                                                                   | Discovery mode (`recommendation` or deprecated `legacy`). Legacy mode no longer receives fixes and will be removed in a future release; migrate by unsetting `DISCOVERY_MODE`.                                                                                                                                                                                                                                         |
 | `AUDIOBOOKSHELF_URL`                     | `backend`, `backend-worker`                    | Optional                                      | unset                                                                              | Audiobookshelf service URL (env fallback path).                                                                                                                                                                                                                                                                                                                                                                      |
 | `AUDIOBOOKSHELF_API_KEY`                 | `backend`, `backend-worker`                    | Optional (required if using the env fallback) | unset                                                                              | Audiobookshelf API key for env-based fallback configuration. Consumed with `AUDIOBOOKSHELF_URL` via `config.audiobookshelf` when no database-stored Audiobookshelf settings exist.                                                                                                                                                                                                                                   |
 | `TIDAL_SIDECAR_URL`                      | `backend`                                      | Optional                                      | split stack: `http://tidal-downloader:8585`; AIO/host-run: `http://127.0.0.1:8585` | URL for TIDAL sidecar service.                                                                                                                                                                                                                                                                                                                                                                                       |
