+++ docs/modernization-roadmap.md
@@ -470,15 +470,15 @@ _(includes F54; dimension tallies double-count the F2/F44 and F17/F47 duplicate 
 
 **⬜ open** · dimension: idempotency · severity: high→medium · effort: M · risk: low · epic: #13
 
-> **Audit note.** ✓ Severity high→medium: the missing guard is confined to the non-default `DISCOVERY_MODE=legacy` path. NOTE: F19's shipped queue-wide `attempts:3` retry now makes this load-bearing — a legacy retry can duplicate the batch and re-trigger Lidarr downloads.
+> **Audit note.** ✓ Severity high→medium: the missing guard is confined to the deprecated, non-default `DISCOVERY_MODE=legacy` path. The legacy route is frozen and will be removed with the mode, so this duplicate-retry hazard remains intentionally unfixed. NOTE: F19's shipped queue-wide `attempts:3` retry makes the hazard load-bearing — a legacy retry can duplicate the batch and re-trigger Lidarr downloads.
 
-**Files:** `backend/src/services/discoverWeekly.ts:329`, `backend/src/services/discoverWeekly.ts:331`, `backend/src/routes/discover.ts:217`, `backend/prisma/schema.prisma`
+**Files:** `backend/src/services/discoverWeekly.ts:329`, `backend/src/services/discoverWeekly.ts:331`, `backend/src/routes/discover/legacy/generation.ts`, `backend/prisma/schema.prisma`
 
-**Problem.** generatePlaylist computes weekStart (discoverWeekly.ts:205) but the service NEVER checks for an existing DiscoveryBatch for (userId, weekStart) before tx.discoveryBatch.create (line 331) — a confirmed grep over lines 196-330 found no findFirst/findUnique/upsert on discoveryBatch. The only existence guard is a 409 in the HTTP route (discover.ts:198-212), which does NOT protect the cron path, the legacy discoverQueue.add({ userId }) at discover.ts:217 (no jobId, so no Bull dedup either), a Bull retry, or a double-fired job. Two batches in the same week double the album set sent to Lidarr, doubling bandwidth and download quota. DiscoveryBatch has only @@index([userId, weekStart]) — no @@unique — so even concurrent creates aren't blocked at the DB.
+**Problem.** generatePlaylist computes weekStart (discoverWeekly.ts:205) but the service NEVER checks for an existing DiscoveryBatch for (userId, weekStart) before tx.discoveryBatch.create (line 331) — a confirmed grep over lines 196-330 found no findFirst/findUnique/upsert on discoveryBatch. The only existence guard is a 409 in the frozen legacy HTTP handler (`backend/src/routes/discover/legacy/generation.ts`), which does NOT protect the cron path, its legacy `discoverQueue.add({ userId })` call (no jobId, so no Bull dedup either), a Bull retry, or a double-fired job. Two batches in the same week double the album set sent to Lidarr, doubling bandwidth and download quota. DiscoveryBatch has only @@index([userId, weekStart]) — no @@unique — so even concurrent creates aren't blocked at the DB.
 
-**Recommendation.** Scope the fix to legacy mode (the only path that creates DiscoveryBatch + Lidarr downloads); the default 'recommendation' path is already idempotent (deletes prior-week discoveryAlbum/Track before recreating). Several of F20's claimed gaps are already closed: (1) the cron path IS deduped via a deterministic week-scoped Bull jobId discover:cron:${weekKey}:${userId} (discoverCron.ts:17-60) — F20's 'cron unprotected' claim is wrong; (2) a per-user Redis SET NX lock (discoverProcessor.ts:107-144) serializes concurrent jobs, so simultaneous creates can't both fire; (3) per-album downloadJob.findFirst dedup (discoverWeekly.ts:360-373) prevents re-downloading albums with an in-flight job. The genuine residual gap: a COMPLETED batch re-triggered the same week (manual legacy route line 217 enqueues with NO jobId, or a Bull retry of the unnamed job) creates a duplicate DiscoveryBatch — the route 409 only blocks downloading/scanning status and generatePlaylist has no terminal-batch check. Minimal fix: (a) at top of discoverWeekly.ts generatePlaylist add a findFirst on discoveryBatch for (userId, weekStart) and short-circuit if one exists in any non-failed/non-error state, returning the existing batchId; (b) give the legacy route's discoverQueue.add({ userId }) a deterministic jobId so duplicate manual enqueues coalesce like the cron path already does. Defer the @@unique/partial-index DB constraint unless you first audit+backfill existing duplicate (userId, weekStart) rows; if added, use a status-scoped partial unique index via a Prisma migration rather than a blanket @@unique, and update CHANGELOG.md.
+**Recommendation.** Do not implement the earlier fix proposal. The maintainer lifecycle decision freezes legacy discovery without further fixes and removes this hazard with the mode. The default `recommendation` path is already idempotent. Historical analysis: the cron path is deduped via a deterministic week-scoped Bull jobId `discover:cron:${weekKey}:${userId}`; a per-user Redis SET NX lock serializes concurrent jobs; and per-album downloadJob.findFirst dedup prevents re-downloading albums with an in-flight job. The residual gap is a COMPLETED legacy batch re-triggered in the same week by the unnamed manual job or a Bull retry.
 
-**Safety / pitfalls.** Adding @@unique([userId, weekStart]) to DiscoveryBatch is NOT safe as a plain migration: legacy mode's HTTP route (discover.ts:217) and a fully-completed-then-re-triggered batch can legitimately produce a second row in the same week today, so the existing prod table may already contain duplicate (userId, weekStart) rows — the migration would fail mid-deploy unless preceded by a dedup/backfill step. A status-scoped partial unique index (active batches only) is safer but requires raw SQL via a Prisma migration, which sits at the edge of the AGENTS.md \"no raw SQL\" contract (migrations are the accepted exception, but call it out). Switching the bare create to an upsert on (userId, weekStart) would also silently overwrite/merge a prior week's batch logs/counters if the unique key collides with an older completed batch, changing observable history. Note also the default code path (DISCOVERY_MODE=recommendation) does not use DiscoveryBatch at all and is already week-idempotent, so this change only affects the non-default legacy path.
+**Safety / pitfalls.** Adding @@unique([userId, weekStart]) to DiscoveryBatch is NOT safe as a plain migration: legacy mode's frozen HTTP handler and a fully-completed-then-re-triggered batch can legitimately produce a second row in the same week today, so the existing prod table may already contain duplicate (userId, weekStart) rows — the migration would fail mid-deploy unless preceded by a dedup/backfill step. A status-scoped partial unique index (active batches only) is safer but requires raw SQL via a Prisma migration, which sits at the edge of the AGENTS.md \"no raw SQL\" contract (migrations are the accepted exception, but call it out). Switching the bare create to an upsert on (userId, weekStart) would also silently overwrite/merge a prior week's batch logs/counters if the unique key collides with an older completed batch, changing observable history. Note also the default code path (DISCOVERY_MODE=recommendation) does not use DiscoveryBatch at all and is already week-idempotent, so this change only affects the deprecated legacy path.
 
 ---
