+++ scripts/ci/check-file-size.mjs
@@ -55,7 +55,6 @@ const FIXED_SCAN_ROOTS = Object.freeze([
 const BASELINE = Object.freeze({
     "backend/src/routes/audiobooks.ts": 1555,
     "backend/src/routes/browse.ts": 1557,
-    "backend/src/routes/discover.ts": 2429,
     "backend/src/routes/downloads.ts": 1766,
     "backend/src/routes/enrichment.ts": 2277,
     "backend/src/routes/library/artists.ts": 1553,
